backend/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
gunicorn==21.2.0
mysqlclient==2.2.4